+++ mopidy/__main__.py
@@ -1,7 +1,6 @@
 from __future__ import unicode_literals
 
 import logging
-import optparse
 import os
 import signal
 import sys
@@ -18,17 +17,11 @@ mopidy_args = sys.argv[1:]
 sys.argv[1:] = []
 
 
-# Add ../ to the path so we can run Mopidy from a Git checkout without
-# installing it on the system.
-sys.path.insert(
-    0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../')))
-
-
-from mopidy import ext
+from mopidy import commands, ext
 from mopidy.audio import Audio
 from mopidy import config as config_lib
 from mopidy.core import Core
-from mopidy.utils import deps, log, path, process, versioning
+from mopidy.utils import log, path, process
 
 logger = logging.getLogger('mopidy.main')
 
@@ -37,33 +30,36 @@ def main():
     signal.signal(signal.SIGTERM, process.exit_handler)
     signal.signal(signal.SIGUSR1, pykka.debug.log_thread_tracebacks)
 
-    loop = gobject.MainLoop()
-    options = parse_options()
-    config_files = options.config.split(b':')
-    config_overrides = options.overrides
-
-    enabled_extensions = []  # Make sure it is defined before the finally block
-    logging_initialized = False
+    args = commands.parser.parse_args(args=mopidy_args)
+    if args.show_config:
+        commands.show_config(args)
+    if args.show_deps:
+        commands.show_deps()
 
     # TODO: figure out a way to make the boilerplate in this file reusable in
     # scanner and other places we need it.
 
     try:
         # Initial config without extensions to bootstrap logging.
-        logging_config, _ = config_lib.load(config_files, [], config_overrides)
+        logging_initialized = False
+        logging_config, _ = config_lib.load(
+            args.config_files, [], args.config_overrides)
 
         # TODO: setup_logging needs defaults in-case config values are None
         log.setup_logging(
-            logging_config, options.verbosity_level, options.save_debug_log)
+            logging_config, args.verbosity_level, args.save_debug_log)
         logging_initialized = True
+
+        create_file_structures()
+        check_old_locations()
 
         installed_extensions = ext.load_extensions()
 
-        # TODO: wrap config in RO proxy.
         config, config_errors = config_lib.load(
-            config_files, installed_extensions, config_overrides)
+            args.config_files, installed_extensions, args.config_overrides)
 
         # Filter out disabled extensions and remove any config errors for them.
+        enabled_extensions = []
         for extension in installed_extensions:
             enabled = config[extension.ext_name]['enabled']
             if ext.validate_extension(extension) and enabled:
@@ -78,31 +74,38 @@ def main():
         proxied_config = config_lib.Proxy(config)
 
         log.setup_log_levels(proxied_config)
-        create_file_structures()
-        check_old_locations()
         ext.register_gstreamer_elements(enabled_extensions)
 
         # Anything that wants to exit after this point must use
         # mopidy.utils.process.exit_process as actors have been started.
-        audio = setup_audio(proxied_config)
-        backends = setup_backends(proxied_config, enabled_extensions, audio)
-        core = setup_core(audio, backends)
-        setup_frontends(proxied_config, enabled_extensions, core)
-        loop.run()
+        start(proxied_config, enabled_extensions)
     except KeyboardInterrupt:
-        if logging_initialized:
-            logger.info('Interrupted. Exiting...')
+        pass
     except Exception as ex:
         if logging_initialized:
             logger.exception(ex)
         raise
-    finally:
-        loop.quit()
-        stop_frontends(enabled_extensions)
-        stop_core()
-        stop_backends(enabled_extensions)
-        stop_audio()
-        process.stop_remaining_actors()
+
+
+def create_file_structures():
+    path.get_or_create_dir(b'$XDG_DATA_DIR/mopidy')
+    path.get_or_create_file(b'$XDG_CONFIG_DIR/mopidy/mopidy.conf')
+
+
+def check_old_locations():
+    dot_mopidy_dir = path.expand_path(b'~/.mopidy')
+    if os.path.isdir(dot_mopidy_dir):
+        logger.warning(
+            'Old Mopidy dot dir found at %s. Please migrate your config to '
+            'the ini-file based config format. See release notes for further '
+            'instructions.', dot_mopidy_dir)
+
+    old_settings_file = path.expand_path(b'$XDG_CONFIG_DIR/mopidy/settings.py')
+    if os.path.isfile(old_settings_file):
+        logger.warning(
+            'Old Mopidy settings file found at %s. Please migrate your '
+            'config to the ini-file based config format. See release notes '
+            'for further instructions.', old_settings_file)
 
 
 def log_extension_info(all_extensions, enabled_extensions):
@@ -124,102 +127,27 @@ def check_config_errors(errors):
     sys.exit(1)
 
 
-def check_config_override(option, opt, override):
+def start(config, extensions):
+    loop = gobject.MainLoop()
     try:
-        return config_lib.parse_override(override)
-    except ValueError:
-        raise optparse.OptionValueError(
-            'option %s: must have the format section/key=value' % opt)
+        audio = start_audio(config)
+        backends = start_backends(config, extensions, audio)
+        core = start_core(audio, backends)
+        start_frontends(config, extensions, core)
+        loop.run()
+    except KeyboardInterrupt:
+        logger.info('Interrupted. Exiting...')
+        return
+    finally:
+        loop.quit()
+        stop_frontends(extensions)
+        stop_core()
+        stop_backends(extensions)
+        stop_audio()
+        process.stop_remaining_actors()
 
 
-def parse_options():
-    parser = optparse.OptionParser(
-        version='Mopidy %s' % versioning.get_version())
-
-    # Ugly extension of optparse type checking magic :/
-    optparse.Option.TYPES += ('config_override',)
-    optparse.Option.TYPE_CHECKER['config_override'] = check_config_override
-
-    # NOTE First argument to add_option must be bytestrings on Python < 2.6.2
-    # See https://github.com/mopidy/mopidy/issues/302 for details
-    parser.add_option(
-        b'-q', '--quiet',
-        action='store_const', const=0, dest='verbosity_level',
-        help='less output (warning level)')
-    parser.add_option(
-        b'-v', '--verbose',
-        action='count', default=1, dest='verbosity_level',
-        help='more output (debug level)')
-    parser.add_option(
-        b'--save-debug-log',
-        action='store_true', dest='save_debug_log',
-        help='save debug log to "./mopidy.log"')
-    parser.add_option(
-        b'--show-config',
-        action='callback', callback=show_config_callback,
-        help='show current config')
-    parser.add_option(
-        b'--show-deps',
-        action='callback', callback=deps.show_deps_optparse_callback,
-        help='show dependencies and their versions')
-    parser.add_option(
-        b'--config',
-        action='store', dest='config',
-        default=b'$XDG_CONFIG_DIR/mopidy/mopidy.conf',
-        help='config files to use, colon seperated, later files override')
-    parser.add_option(
-        b'-o', b'--option',
-        action='append', dest='overrides', type='config_override',
-        help='`section/key=value` values to override config options')
-    return parser.parse_args(args=mopidy_args)[0]
-
-
-def show_config_callback(option, opt, value, parser):
-    # TODO: don't use callback for this as --config or -o set after
-    # --show-config will be ignored.
-    files = getattr(parser.values, 'config', b'').split(b':')
-    overrides = getattr(parser.values, 'overrides', [])
-
-    extensions = ext.load_extensions()
-    config, errors = config_lib.load(files, extensions, overrides)
-
-    # Clear out any config for disabled extensions.
-    for extension in extensions:
-        if not ext.validate_extension(extension):
-            config[extension.ext_name] = {b'enabled': False}
-            errors[extension.ext_name] = {
-                b'enabled': b'extension disabled its self.'}
-        elif not config[extension.ext_name]['enabled']:
-            config[extension.ext_name] = {b'enabled': False}
-            errors[extension.ext_name] = {
-                b'enabled': b'extension disabled by config.'}
-
-    print config_lib.format(config, extensions, errors)
-    sys.exit(0)
-
-
-def check_old_locations():
-    dot_mopidy_dir = path.expand_path(b'~/.mopidy')
-    if os.path.isdir(dot_mopidy_dir):
-        logger.warning(
-            'Old Mopidy dot dir found at %s. Please migrate your config to '
-            'the ini-file based config format. See release notes for further '
-            'instructions.', dot_mopidy_dir)
-
-    old_settings_file = path.expand_path(b'$XDG_CONFIG_DIR/mopidy/settings.py')
-    if os.path.isfile(old_settings_file):
-        logger.warning(
-            'Old Mopidy settings file found at %s. Please migrate your '
-            'config to the ini-file based config format. See release notes '
-            'for further instructions.', old_settings_file)
-
-
-def create_file_structures():
-    path.get_or_create_dir(b'$XDG_DATA_DIR/mopidy')
-    path.get_or_create_file(b'$XDG_CONFIG_DIR/mopidy/mopidy.conf')
-
-
-def setup_audio(config):
+def start_audio(config):
     logger.info('Starting Mopidy audio')
     return Audio.start(config=config).proxy()
 
@@ -229,7 +157,7 @@ def stop_audio():
     process.stop_actors_by_class(Audio)
 
 
-def setup_backends(config, extensions, audio):
+def start_backends(config, extensions, audio):
     backend_classes = []
     for extension in extensions:
         backend_classes.extend(extension.get_backend_classes())
@@ -253,7 +181,7 @@ def stop_backends(extensions):
             process.stop_actors_by_class(backend_class)
 
 
-def setup_core(audio, backends):
+def start_core(audio, backends):
     logger.info('Starting Mopidy core')
     return Core.start(audio=audio, backends=backends).proxy()
 
@@ -263,7 +191,7 @@ def stop_core():
     process.stop_actors_by_class(Core)
 
 
-def setup_frontends(config, extensions, core):
+def start_frontends(config, extensions, core):
     frontend_classes = []
     for extension in extensions:
         frontend_classes.extend(extension.get_frontend_classes())