+++ mopidy/backends/local/__init__.py
@@ -21,6 +21,7 @@ class Extension(ext.Extension):
         schema['media_dir'] = config.Path()
         schema['playlists_dir'] = config.Path()
         schema['tag_cache_file'] = config.Path()
+        schema['scan_timeout'] = config.Integer(minimum=0)
         return schema
 
     def validate_environment(self):
@@ -29,3 +30,7 @@ class Extension(ext.Extension):
     def get_backend_classes(self):
         from .actor import LocalBackend
         return [LocalBackend]
+
+    def get_library_updaters(self):
+        from .library import LocalLibraryUpdateProvider
+        return [LocalLibraryUpdateProvider]