+++ tests/frontends/mpd/protocol/music_db_test.py
@@ -7,7 +7,19 @@ from tests.frontends.mpd import protocol
 
 class MusicDatabaseHandlerTest(protocol.BaseTestCase):
     def test_count(self):
-        self.sendRequest('count "tag" "needle"')
+        self.sendRequest('count "artist" "needle"')
+        self.assertInResponse('songs: 0')
+        self.assertInResponse('playtime: 0')
+        self.assertInResponse('OK')
+
+    def test_count_without_quotes(self):
+        self.sendRequest('count artist "needle"')
+        self.assertInResponse('songs: 0')
+        self.assertInResponse('playtime: 0')
+        self.assertInResponse('OK')
+
+    def test_count_with_multiple_pairs(self):
+        self.sendRequest('count "artist" "foo" "album" "bar"')
         self.assertInResponse('songs: 0')
         self.assertInResponse('playtime: 0')
         self.assertInResponse('OK')
@@ -70,11 +82,19 @@ class MusicDatabaseHandlerTest(protocol.BaseTestCase):
         self.assertEqual(playlists[0].tracks[0].uri, 'dummy:a')
         self.assertInResponse('OK')
 
-    def test_listall(self):
+    def test_listall_without_uri(self):
+        self.sendRequest('listall')
+        self.assertEqualResponse('ACK [0@0] {} Not implemented')
+
+    def test_listall_with_uri(self):
         self.sendRequest('listall "file:///dev/urandom"')
         self.assertEqualResponse('ACK [0@0] {} Not implemented')
 
-    def test_listallinfo(self):
+    def test_listallinfo_without_uri(self):
+        self.sendRequest('listallinfo')
+        self.assertEqualResponse('ACK [0@0] {} Not implemented')
+
+    def test_listallinfo_with_uri(self):
         self.sendRequest('listallinfo "file:///dev/urandom"')
         self.assertEqualResponse('ACK [0@0] {} Not implemented')