+++ tests/frontends/mpd/protocol/stored_playlists_test.py
@@ -10,8 +10,8 @@ from tests.frontends.mpd import protocol
 class PlaylistsHandlerTest(protocol.BaseTestCase):
     def test_listplaylist(self):
         self.backend.playlists.playlists = [
-            Playlist(name='name', uri='dummy:name',
-                tracks=[Track(uri='dummy:a')])]
+            Playlist(
+                name='name', uri='dummy:name', tracks=[Track(uri='dummy:a')])]
 
         self.sendRequest('listplaylist "name"')
         self.assertInResponse('file: dummy:a')
@@ -19,8 +19,8 @@ class PlaylistsHandlerTest(protocol.BaseTestCase):
 
     def test_listplaylist_without_quotes(self):
         self.backend.playlists.playlists = [
-            Playlist(name='name', uri='dummy:name', 
-                tracks=[Track(uri='dummy:a')])]
+            Playlist(
+                name='name', uri='dummy:name', tracks=[Track(uri='dummy:a')])]
 
         self.sendRequest('listplaylist name')
         self.assertInResponse('file: dummy:a')
@@ -41,8 +41,8 @@ class PlaylistsHandlerTest(protocol.BaseTestCase):
 
     def test_listplaylistinfo(self):
         self.backend.playlists.playlists = [
-            Playlist(name='name', uri='dummy:name', 
-                tracks=[Track(uri='dummy:a')])]
+            Playlist(
+                name='name', uri='dummy:name', tracks=[Track(uri='dummy:a')])]
 
         self.sendRequest('listplaylistinfo "name"')
         self.assertInResponse('file: dummy:a')
@@ -52,8 +52,8 @@ class PlaylistsHandlerTest(protocol.BaseTestCase):
 
     def test_listplaylistinfo_without_quotes(self):
         self.backend.playlists.playlists = [
-            Playlist(name='name', uri='dummy:name', 
-                tracks=[Track(uri='dummy:a')])]
+            Playlist(
+                name='name', uri='dummy:name', tracks=[Track(uri='dummy:a')])]
 
         self.sendRequest('listplaylistinfo name')
         self.assertInResponse('file: dummy:a')
@@ -105,6 +105,30 @@ class PlaylistsHandlerTest(protocol.BaseTestCase):
 
         self.sendRequest('listplaylists')
         self.assertNotInResponse('playlist: ')
+        self.assertInResponse('OK')
+
+    def test_listplaylists_replaces_newline_with_space(self):
+        self.backend.playlists.playlists = [
+            Playlist(name='a\n', uri='dummy:')]
+        self.sendRequest('listplaylists')
+        self.assertInResponse('playlist: a ')
+        self.assertNotInResponse('playlist: a\n')
+        self.assertInResponse('OK')
+
+    def test_listplaylists_replaces_carriage_return_with_space(self):
+        self.backend.playlists.playlists = [
+            Playlist(name='a\r', uri='dummy:')]
+        self.sendRequest('listplaylists')
+        self.assertInResponse('playlist: a ')
+        self.assertNotInResponse('playlist: a\r')
+        self.assertInResponse('OK')
+
+    def test_listplaylists_replaces_forward_slash_with_space(self):
+        self.backend.playlists.playlists = [
+            Playlist(name='a/', uri='dummy:')]
+        self.sendRequest('listplaylists')
+        self.assertInResponse('playlist: a ')
+        self.assertNotInResponse('playlist: a/')
         self.assertInResponse('OK')
 
     def test_load_appends_to_tracklist(self):